*** empty log message ***
git-svn-id: http://svn.vuze.com/public/client@2156 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/download/DownloadManagerStats.java
@@ -57,13 +57,13 @@ DownloadManagerStats
 	public int
 	getShareRatio();
 	
-	public int
+	public long
 	getDownloadAverage();
 		
-	public int
+	public long
 	getUploadAverage();
 
-	public int
+	public long
 	getTotalAverage();
 			
 	public String
